+++ js/app/components/RepositoryList.react.js
@@ -2,69 +2,19 @@ var
   React = require( "react" ),
   Link = require("react-router").Link,
 
-  RepositoryListActions = require( "../actions/RepositoryListActions" ),
+  RepositoryActions = require( "../actions/RepositoryActions" ),
   UserStore = require( "../stores/UserStore" ),
-  RepositoryListStore = require( "../stores/RepositoryListStore" ),
+  RepositoryStore = require( "../stores/RepositoryStore" ),
   AppHeader = require( "./AppHeader.react" ),
   Breadcrumbs = require( "./Breadcrumbs.react" ),
   RepositoryListHead = require( "./RepositoryListHead.react" ),
-  RepositoryListItem = require( "./RepositoryListItem.react" );
+  RepositoryListItem = require( "./RepositoryListItem.react" ),
 
   getRepositoryListState = function() {
     return {
       token: UserStore.getToken(),
       user: UserStore.getUser(),
-      repositories: RepositoryListStore.getRepositoryList()
-      //repositories: [
-      //  {
-      //    id: 1,
-      //    name: "Kanbaner",
-      //    full_name: "zapix/kanbaner",
-      //    open_issues_count: 12,
-      //    updated_at: "2011-01-26T19:14:43Z",
-      //    has_issues: true,
-      //    owner: {
-      //      id: 346813,
-      //      login: "Zapix"
-      //    }
-      //  },
-      //  {
-      //    id: 4,
-      //    name: "Oblichi-content",
-      //    full_name: "oblichi/oblichi-content",
-      //    open_issues_count: 5,
-      //    updated_at: "2011-01-26T19:14:43Z",
-      //    has_issues: true,
-      //    owner: {
-      //      id: 43,
-      //      login: "atorich"
-      //    }
-      //  },
-      //  {
-      //    id: 8,
-      //    name: "tmnv1",
-      //    full_name: "dtmnv/tmnv1",
-      //    open_issues_count: 10,
-      //    updadated_at: "2011-01-26T19:14:43Z",
-      //    has_issues: true,
-      //    owner: {
-      //      id: 16,
-      //      login: "dtmnv"
-      //    }
-      //  },
-      //  {
-      //    id: 1,
-      //    name: "django-confirmaction",
-      //    full_name: "zapix/django-confirmcation",
-      //    open_issues_count: 0,
-      //    updated_at: "2011-01-26T19:14:43Z",
-      //    has_issues: false,
-      //    owner: {
-      //      id: 346813,
-      //      login: "Zapix"
-      //    }
-      //  }
-      //]
+      repositories: RepositoryStore.getRepositoryList()
     }
 
   },
@@ -81,12 +31,12 @@ var
      * info about repositories
      */
     componentDidMount: function() {
-      RepositoryListStore.addRepositoryListChangedListener(
+      RepositoryStore.addRepositoryListChangedListener(
         this.onRepositoryListChanged
       );
 
       if( this.state.repositories.length == 0) {
-        RepositoryListActions.requestRepositoryList(this.state.token);
+        RepositoryActions.requestRepositoryList(this.state.token);
       }
     },
 
@@ -94,7 +44,7 @@ var
      * Unsubscribe form REPOSITORY_LIST_CHANGED event
      */
     componentWillUnmount: function() {
-      RepositoryListStore.removeRepositoryListChangedListener(
+      RepositoryStore.removeRepositoryListChangedListener(
         this.onRepositoryListChanged
       );
     },